feat: Add conditional rendering for no bookmarks in bookmark page
This commit adds conditional rendering to the bookmark page in order to display a message when there are no bookmarks. The code checks the length of the data array and renders a "No Bookmarks" message if it is empty. This enhancement improves the user experience by providing clear feedback when there are no bookmarked items.